src/requirements.txt
absl-py==1.0.0
black==22.3.0
chex==0.1.3
click==8.1.3
cycler==0.11.0
dm-tree==0.1.7
flatbuffers==2.0
flax==0.4.2
fonttools==4.33.3
gurobipy==9.5.1
jax==0.3.10
jaxlib==0.3.10
kiwisolver==1.4.2
matplotlib==3.5.2
mypy-extensions==0.4.3
numpy==1.22.3
opt-einsum==3.3.0
optax==0.1.2
packaging==21.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
pyparsing==3.0.9
python-dateutil==2.8.2
scipy==1.8.0
setuptools==49.2.1
six==1.16.0
tomli==2.0.1
toolz==0.11.2
typing-extensions==4.2.0
